# zema_store_backend/src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zema Store API with Swagger",
        "version": "1.0.0",
        "description": "Zema Store API",
        "contact": {
            "name": "Zema Store",
            "url": "http://zema-store.herokuapp.com",
            "email": "henokadane8@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://zema-store.herokuapp.com/api"
        },
        {
            "url": "http://localhost:3000/api"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/songs/album/{albumId}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Returns list of songs",
                "description": "Returns list of songs that belong to an album",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "albumId",
                        "required": true,
                        "description": "id of the specified album"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of song items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "song object field to sort returned songs",
                        "enum": [
                            "title",
                            "releaseDate",
                            "createdAt"
                        ],
                        "example": "title:asc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Returns album data",
                "description": "Returns album data specified by id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified album"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Albums"
                ],
                "summary": "Updates album data",
                "description": "Updates album data with specified id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified album"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Album"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Deletes album data",
                "description": "Deletes album data with specified id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified album"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Returns list of albums",
                "description": "Returns list of all albums with pagination",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of album items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "album object field to sort returned albums",
                        "enum": [
                            "title",
                            "createdAt"
                        ],
                        "example": "title:asc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Adds album",
                "description": "Adds album",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Album"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuthorization": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token should be accessToken for user"
            }
        },
...
        "schemas": {
            "SuccessMessage": {
                "type": "string",
                "description": "Success message",
                "nullable": true,
                "required": true
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "required": true
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": true
                    }
                }
            },
            "OkResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/SuccessMessage"
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": false
                    }
                },
                "required": [
                    "message",
                    "data"
                ]
            },
...
            "Album": {
                "type": "object",
                "properties": {
                    "artistId": {
                        "type": "string",
                        "description": "Artist Id"
                    },
                    "title": {
                        "type": "string",
                        "description": "Album title"
                    },
                    "photo": {
                        "type": "string",
                        "format": "binary",
                        "description": "Album photo"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Album release date",
                        "example": "2022-07-21"
                    }
                }
            },
...
        }
    }
}